refactor: added util for generating typescript typing code based on graphql schema, and general code improvements
