+++ README.md
@@ -8,11 +8,11 @@ It uses API from https://corona.lmao.ninja/ and converted it into GraphQL API.
 
 1. Clone the repository and run `yarn install` to install the dependency
 
-2. Run `yarn dev` to run the dev server
+2. Run `yarn start` to run the dev server
 
 3. Use your faviourite API tool to interact with the API
 
-
+ps it uses graphql-codegen for typing code generation (Run `yarn generate:graphql` to generate code it is also a part of `yarn start` )
 ## License
 This project is not licensed for free use. And all right reserved to Les Mills International.
