+++ graphql-codegen.yml
@@ -1,0 +1,8 @@
+overwrite: true
+schema: 'src/schemas/country.graphql'
+documents: null
+generates:
+  generated/graphql.ts:
+    plugins:
+      - 'typescript'
+      - 'typescript-resolvers'
