+++ package.json
@@ -4,7 +4,9 @@
   "main": "index.js",
   "license": "UNLICENSED",
   "scripts": {
-    "dev": "nodemon --config nodemon.json src/server.ts"
+    "generate:graphql": "graphql-codegen --config graphql-codegen.yml",
+    "dev": "nodemon --config nodemon.json src/server.ts",
+    "start": "yarn generate:graphql && yarn dev"
   },
   "dependencies": {
     "@graphql-tools/load-files": "^6.5.3",
@@ -15,6 +17,9 @@
     "lodash": "^4.17.21"
   },
   "devDependencies": {
+    "@graphql-codegen/cli": "^2.6.2",
+    "@graphql-codegen/typescript": "^2.4.5",
+    "@graphql-codegen/typescript-resolvers": "^2.5.2",
     "@types/lodash": "^4.14.179",
     "apollo-datasource-rest": "^3.5.1",
     "nodemon": "^2.0.15",
